Modification des types d'identifiant de Long à Integer dans les dépôts et mise à jour des tests associés

+++ back/src/main/java/com/openclassrooms/ycyw/repository/ChatRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.openclassrooms.ycyw.models.Chat;
 
 @Repository
-public interface ChatRepository extends JpaRepository<Chat, Long> {
+public interface ChatRepository extends JpaRepository<Chat, Integer> {
 }
